+++ pyproject.toml
@@ -85,7 +85,7 @@ all = [
 
 [tool.setuptools]
 package-dir = {"" = "src"}
-packages = ["mxcp", "mxcp.cli", "mxcp.config", "mxcp.config.config_schemas", "mxcp.dbt", "mxcp.drift", "mxcp.drift.drift_schemas", "mxcp.endpoints", "mxcp.endpoints.endpoint_schemas", "mxcp.evals", "mxcp.evals.eval_schemas", "mxcp.plugins", "mxcp.policy", "mxcp.runtime", "mxcp.sdk", "mxcp.sdk.audit", "mxcp.sdk.auth", "mxcp.sdk.core", "mxcp.sdk.core.analytics", "mxcp.sdk.core.config", "mxcp.sdk.core.config.resolvers", "mxcp.sdk.evals", "mxcp.sdk.executor", "mxcp.sdk.executor.plugins", "mxcp.sdk.executor.plugins.duckdb_plugin", "mxcp.sdk.executor.plugins.python_plugin", "mxcp.sdk.policy", "mxcp.sdk.validator", "mxcp.validator"]
+packages = ["mxcp", "mxcp.cli", "mxcp.config", "mxcp.config.config_schemas", "mxcp.dbt", "mxcp.drift", "mxcp.drift.drift_schemas", "mxcp.endpoints", "mxcp.endpoints.endpoint_schemas", "mxcp.evals", "mxcp.evals.eval_schemas", "mxcp.plugins", "mxcp.policy", "mxcp.runtime", "mxcp.sdk", "mxcp.sdk.audit", "mxcp.sdk.auth", "mxcp.sdk.auth.providers", "mxcp.sdk.core", "mxcp.sdk.core.analytics", "mxcp.sdk.core.config", "mxcp.sdk.core.config.resolvers", "mxcp.sdk.evals", "mxcp.sdk.executor", "mxcp.sdk.executor.plugins", "mxcp.sdk.executor.plugins.duckdb_plugin", "mxcp.sdk.executor.plugins.python_plugin", "mxcp.sdk.policy", "mxcp.sdk.validator", "mxcp.sdk.validator.schemas"]
 include-package-data = true
 
 [tool.setuptools.package-data]
@@ -95,7 +95,7 @@ include-package-data = true
 "mxcp.drift.drift_schemas" = ["*.json"]
 "mxcp.evals.eval_schemas" = ["*.json"]
 "mxcp.sdk.core.config.schemas" = ["*.json"]
-"mxcp.validator.schemas" = ["*.json"]
+"mxcp.sdk.validator.schemas" = ["*.json"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
