+++ src/mxcp/sdk/auth/__init__.py
@@ -72,7 +72,7 @@ from ._types import (
 )
 from .context import get_user_context, reset_user_context, set_user_context
 from .middleware import AuthenticationMiddleware
-from .providers import (
+from .base import (
     ExternalOAuthHandler,
     GeneralOAuthAuthorizationServer,
 )
