+++ src/mxcp/sdk/auth/base.py
