+++ src/mxcp/sdk/auth/providers/github.py
@@ -1,4 +1,4 @@
-"""Salesforce OAuth provider implementation for MXCP authentication."""
+"""GitHub OAuth provider implementation for MXCP authentication."""
 
 import logging
 import secrets
@@ -10,51 +10,40 @@ from starlette.exceptions import HTTPException
 from starlette.requests import Request
 from starlette.responses import HTMLResponse, RedirectResponse, Response
 
-from ._types import (
-    ExternalUserInfo,
-    HttpTransportConfig,
-    SalesforceAuthConfig,
-    StateMeta,
-    UserContext,
-)
-from .providers import ExternalOAuthHandler, GeneralOAuthAuthorizationServer
-from .url_utils import URLBuilder
+from .._types import ExternalUserInfo, GitHubAuthConfig, HttpTransportConfig, StateMeta, UserContext
+from ..base import ExternalOAuthHandler, GeneralOAuthAuthorizationServer
+from ..url_utils import URLBuilder
 
 logger = logging.getLogger(__name__)
 
 
-class SalesforceOAuthHandler(ExternalOAuthHandler):
-    """Salesforce OAuth provider implementation for Salesforce Cloud."""
+class GitHubOAuthHandler(ExternalOAuthHandler):
+    """GitHub OAuth provider implementation."""
 
     def __init__(
         self,
-        salesforce_config: SalesforceAuthConfig,
+        github_config: GitHubAuthConfig,
         transport_config: HttpTransportConfig | None = None,
         host: str = "localhost",
         port: int = 8000,
     ):
-        """Initialize Salesforce OAuth handler.
+        """Initialize GitHub OAuth handler.
 
         Args:
-            salesforce_config: Salesforce-specific OAuth configuration
+            github_config: GitHub-specific OAuth configuration
             transport_config: HTTP transport configuration for URL building
             host: The server host for callback URLs
             port: The server port for callback URLs
         """
-        logger.info(f"SalesforceOAuthHandler init: {salesforce_config}")
+        logger.info(f"GitHubOAuthHandler init: {github_config}")
 
         # Required fields are enforced by TypedDict structure
-        self.client_id = salesforce_config["client_id"]
-        self.client_secret = salesforce_config["client_secret"]
-
-        # Salesforce OAuth endpoints
-        self.auth_url = salesforce_config["auth_url"]
-        self.token_url = salesforce_config["token_url"]
-
-        # Use configured scope or default
-        self.scope = salesforce_config.get("scope", "api refresh_token openid profile email")
-
-        self._callback_path = salesforce_config["callback_path"]
+        self.client_id = github_config["client_id"]
+        self.client_secret = github_config["client_secret"]
+        self.scope = github_config.get("scope", "user:email")
+        self._callback_path = github_config["callback_path"]
+        self.auth_url = github_config["auth_url"]
+        self.token_url = github_config["token_url"]
         self.host = host
         self.port = port
 
@@ -63,6 +52,7 @@ class SalesforceOAuthHandler(ExternalOAuthHandler):
 
         # State storage for OAuth flow
         self._state_store: dict[str, StateMeta] = {}
+        self._callback_store: dict[str, str] = {}  # Store callback URLs separately
 
     # ----- authorize -----
     def get_authorize_url(self, client_id: str, params: AuthorizationParams) -> str:
@@ -79,21 +69,18 @@ class SalesforceOAuthHandler(ExternalOAuthHandler):
             code_challenge=params.code_challenge,
             redirect_uri_provided_explicitly=params.redirect_uri_provided_explicitly,
             client_id=client_id,
-            callback_url=full_callback_url,
         )
+        # Store the callback URL separately for consistency
+        self._callback_store[state] = full_callback_url
 
         logger.info(
-            f"Salesforce OAuth authorize URL: client_id={self.client_id}, redirect_uri={full_callback_url}, scope={self.scope}"
+            f"GitHub OAuth authorize URL: client_id={self.client_id}, redirect_uri={full_callback_url}, scope={self.scope}"
         )
 
-        # Salesforce OAuth parameters
         return (
-            f"{self.auth_url}?"
-            f"client_id={self.client_id}&"
-            f"redirect_uri={full_callback_url}&"
-            f"scope={self.scope}&"
-            f"response_type=code&"
-            f"state={state}"
+            f"{self.auth_url}?client_id={self.client_id}"
+            f"&redirect_uri={full_callback_url}"
+            f"&scope={self.scope}&state={state}"
         )
 
     # ----- state helpers -----
@@ -107,15 +94,16 @@ class SalesforceOAuthHandler(ExternalOAuthHandler):
         self._state_store.pop(state, None)
 
     def cleanup_state(self, state: str) -> None:
-        """Clean up state after OAuth flow completion."""
+        """Clean up state and associated callback URL after OAuth flow completion."""
         self._pop_state(state)
+        self._callback_store.pop(state, None)
 
     # ----- code exchange -----
     async def exchange_code(self, code: str, state: str) -> ExternalUserInfo:
         meta = self.get_state_metadata(state)
 
         # Use the stored callback URL for consistency
-        full_callback_url = meta.callback_url
+        full_callback_url = self._callback_store.get(state)
         if not full_callback_url:
             # Fallback to constructing it using URL builder
             full_callback_url = self.url_builder.build_callback_url(
@@ -123,37 +111,36 @@ class SalesforceOAuthHandler(ExternalOAuthHandler):
             )
 
         logger.info(
-            f"Salesforce OAuth token exchange: code={code[:10]}..., redirect_uri={full_callback_url}"
+            f"GitHub OAuth token exchange: code={code[:10]}..., redirect_uri={full_callback_url}"
         )
 
         async with create_mcp_http_client() as client:
             resp = await client.post(
                 self.token_url,
                 data={
-                    "grant_type": "authorization_code",
                     "client_id": self.client_id,
                     "client_secret": self.client_secret,
                     "code": code,
                     "redirect_uri": full_callback_url,
                 },
-                headers={"Content-Type": "application/x-www-form-urlencoded"},
+                headers={"Accept": "application/json"},
             )
         if resp.status_code != 200:
-            logger.error(f"Salesforce token exchange failed: {resp.status_code} {resp.text}")
+            logger.error(f"GitHub token exchange failed: {resp.status_code} {resp.text}")
             raise HTTPException(400, "Failed to exchange code for token")
         payload = resp.json()
         if "error" in payload:
-            logger.error(f"Salesforce token exchange error: {payload}")
+            logger.error(f"GitHub token exchange error: {payload}")
             raise HTTPException(400, payload.get("error_description", payload["error"]))
 
         # Don't clean up state here - let handle_callback do it after getting metadata
-        logger.info(f"Salesforce OAuth token exchange successful for client: {meta.client_id}")
+        logger.info(f"GitHub OAuth token exchange successful for client: {meta.client_id}")
 
         return ExternalUserInfo(
             id=meta.client_id,
             scopes=[],
             raw_token=payload["access_token"],
-            provider="salesforce",
+            provider="github",
         )
 
     # ----- callback -----
@@ -162,7 +149,7 @@ class SalesforceOAuthHandler(ExternalOAuthHandler):
         return self._callback_path
 
     async def on_callback(
-        self, request: Request, provider: GeneralOAuthAuthorizationServer
+        self, request: Request, provider: "GeneralOAuthAuthorizationServer"
     ) -> Response:  # noqa: E501
         code = request.query_params.get("code")
         state = request.query_params.get("state")
@@ -174,18 +161,18 @@ class SalesforceOAuthHandler(ExternalOAuthHandler):
         except HTTPException:
             raise
         except Exception as exc:
-            logger.error("Salesforce callback failed", exc_info=exc)
+            logger.error("GitHub callback failed", exc_info=exc)
             return HTMLResponse(status_code=500, content="oauth_failure")
 
     # ----- user context -----
     async def get_user_context(self, token: str) -> UserContext:
-        """Get standardized user context from Salesforce.
+        """Get standardized user context from GitHub.
 
         Args:
-            token: Salesforce OAuth access token
+            token: GitHub OAuth access token
 
         Returns:
-            UserContext with Salesforce user information
+            UserContext with GitHub user information
 
         Raises:
             HTTPException: If token is invalid or user info cannot be retrieved
@@ -193,56 +180,28 @@ class SalesforceOAuthHandler(ExternalOAuthHandler):
         try:
             user_profile = await self._fetch_user_profile(token)
 
-            # Extract Salesforce-specific fields and map to standard UserContext
+            # Extract GitHub-specific fields and map to standard UserContext
             return UserContext(
-                provider="salesforce",
-                user_id=str(user_profile.get("user_id", "")),
-                username=user_profile.get(
-                    "username", f"user_{user_profile.get('user_id', 'unknown')}"
-                ),
+                provider="github",
+                user_id=str(user_profile.get("id", "")),
+                username=user_profile.get("login", f"user_{user_profile.get('id', 'unknown')}"),
                 email=user_profile.get("email"),
-                name=user_profile.get("display_name") or user_profile.get("name"),
-                avatar_url=user_profile.get("photos", {}).get("picture"),
+                name=user_profile.get("name"),
+                avatar_url=user_profile.get("avatar_url"),
                 raw_profile=user_profile,
             )
         except Exception as e:
-            logger.error(f"Failed to get Salesforce user context: {e}")
+            logger.error(f"Failed to get GitHub user context: {e}")
             raise HTTPException(500, f"Failed to retrieve user information: {e}") from e
 
     # ----- private helper -----
     async def _fetch_user_profile(self, token: str) -> dict[str, Any]:
-        """Fetch raw user profile from Salesforce UserInfo endpoint (private implementation detail)."""
-        logger.info(f"Fetching Salesforce user profile with token: {token[:10]}...")
-
-        # First get the identity URL from the token response
+        """Fetch raw user profile from GitHub API (private implementation detail)."""
         async with create_mcp_http_client() as client:
-            # Get token info to find the identity URL
-            await client.get(
-                "https://login.salesforce.com/services/oauth2/token",
-                headers={"Authorization": f"Bearer {token}"},
-            )
-
-            # If token info doesn't work, try the standard userinfo endpoint
             resp = await client.get(
-                "https://login.salesforce.com/services/oauth2/userinfo",
+                "https://api.github.com/user",
                 headers={"Authorization": f"Bearer {token}"},
             )
-
-        logger.info(f"Salesforce UserInfo API response: {resp.status_code}")
-
         if resp.status_code != 200:
-            error_body = ""
-            try:
-                error_body = resp.text
-                logger.error(f"Salesforce UserInfo API error {resp.status_code}: {error_body}")
-            except Exception:
-                logger.error(
-                    f"Salesforce UserInfo API error {resp.status_code}: Unable to read response body"
-                )
-            raise ValueError(f"Salesforce API error: {resp.status_code} - {error_body}")
-
-        user_data = cast(dict[str, Any], resp.json())
-        logger.info(
-            f"Successfully fetched Salesforce user profile for user_id: {user_data.get('user_id', 'unknown')}"
-        )
-        return user_data
+            raise ValueError(f"GitHub API error: {resp.status_code}")
+        return cast(dict[str, Any], resp.json())
