+++ src/mxcp/sdk/validator/schemas/validation-schema-1.json
