+++ tests/fixtures/sdk/validator/calculator.yaml
