+++ tests/sdk/validator/test_fixtures.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 
 from mxcp.sdk.validator import TypeValidator, ValidationError
-from mxcp.validator import load_schema_from_file, validate
+from mxcp.sdk.validator import load_schema_from_file, validate
 
 
 class TestFixtures:
@@ -15,7 +15,7 @@ class TestFixtures:
         """Test loading and using calculator schema from fixtures."""
         # Get the fixture path
         fixture_path = (
-            Path(__file__).parent.parent / "fixtures" / "validator" / "schemas" / "calculator.yaml"
+            Path(__file__).parent.parent.parent / "fixtures" / "sdk" / "validator" / "calculator.yaml"
         )
 
         @validate.from_file(str(fixture_path))
@@ -49,7 +49,7 @@ class TestFixtures:
     def test_schema_file_validation(self):
         """Test direct schema loading from file."""
         schema_path = (
-            Path(__file__).parent.parent / "fixtures" / "validator" / "schemas" / "calculator.yaml"
+            Path(__file__).parent.parent.parent / "fixtures" / "sdk" / "validator" / "calculator.yaml"
         )
 
         # Load and create validator
