+++ tests/test_policies_integration.py
@@ -9,7 +9,7 @@ import pytest
 from mxcp.config.site_config import load_site_config
 from mxcp.config.user_config import load_user_config
 from mxcp.endpoints.sdk_executor import execute_endpoint
-from mxcp.sdk.auth.providers import UserContext
+from mxcp.sdk.auth import UserContext
 
 
 @pytest.fixture(scope="session", autouse=True)
