+++ tests/test_tester_assertions.py
@@ -7,7 +7,7 @@ import pytest
 from mxcp.config.site_config import load_site_config
 from mxcp.config.user_config import load_user_config
 from mxcp.endpoints.tester import run_tests
-from mxcp.sdk.auth.providers import UserContext
+from mxcp.sdk.auth import UserContext
 
 
 @pytest.fixture(scope="session", autouse=True)
